WebMP: Create DataResult action result
The DataResult supports returning just a model from the controller, which will
then be either formatted as JSON or rendered as a view, depending on the
request parameters.

This makes it very easy to give skins programmatic access to the underlying
data using AJAX, without needing to create a second action.